pyproject.toml
[tool.poetry]
name = "ui-ux-pro-max-langchain"
version = "0.1.0"
description = "UI/UX Pro Max - LangChain Agent"
authors = ["UI/UX Pro Max Team"]
readme = "README.md"
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
langchain = "^0.2.0"
langchain-openai = "^0.1.0"
langchain-community = "^0.2.0"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
